ai-service/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-dotenv>=1.0.1
openai>=1.30.0
huggingface-hub>=0.23.0
transformers>=4.40.0
torch>=2.2.0
sentence-transformers>=3.0.0
spacy>=3.7.0
numpy>=1.26.0
